Change handling of missing state file
There is no point in getting the user to run `--init`. It's also a bit
dangerous as it might overwrite the contents of an existing file.
Instead, create a new empty state if required. This gets persisted at
the end of the program so the effect is the same as creating the empty
state file.